requirements.txt
pyserial>=3.5
numpy>=1.21
open3d>=0.17